tough: Use u64 for file lengths
u64 is more commonly used across the Rust standard library for file
lengths.

+++ workspaces/updater/tough/src/error.rs
@@ -80,7 +80,7 @@ pub enum Error {
     /// A file's maximum size exceeded a limit set by the consumer of this library or the metadata.
     #[snafu(display("Maximum size {} (specified by {}) exceeded", max_size, specifier))]
     MaxSizeExceeded {
-        max_size: usize,
+        max_size: u64,
         specifier: &'static str,
         backtrace: Backtrace,
     },
